fix: duplicate 608/708 VTT lines
Reuse the hash function when creating the 608/708 VTTCues to assign them a unique id.
Extract the webvtt-parser callback function that adds text track cues and reuse it for adding the 608/708 cues.

Wasn't sure what is the best way to use the hash function - if to leave it in webvtt-aprser and expose or extract it to a seperate vtt utils function and import in both parser and cues modules.
I ended up exposing it from webvtt-parser and adding the id in timeline-controller, but it makes more sense to keep the logic near where VTTCue is created.
Let me know if you want to change this in anyway- open for comments.

In addition, added text field to VTTCue interface.